Add:ボスとプレイヤーのスキルファイル大量追加。追加しただけで、まだ動作はしない。 Update:Geminiによるいくつかの改修

+++ Siv3DGameJam/Siv3DGameJam/DropZone.h
@@ -39,7 +39,7 @@ public:
 
 	void setDropBoxTexture(int32 num, String path, String front = U"");
 
-	String getSkillKey(int slot) const;
+	String getSkillKey(size_t slot) const;
 
 private:
 	std::array<Zone, 4> zones;
